# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CareConnect API Documentation",
        "description": "API documentation for CareConnect Healthcare Management System",
        "version": "1.0.0",
        "contact": {
            "email": "blissbonheur1@gmail.com"
        }
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "API Server V1"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "Error": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "errors": {
                        "type": "object"
                    }
                }
            },
            "Success": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Success message"
                    },
                    "data": {
                        "type": "object"
                    }
                }
            },
...
        }
    },
    "paths": {
...
        "/api/sports/create": {
            "post": {
                "tags": ["Sports"],
                "summary": "Create a new sport recommendation",
                "security": [{"bearerAuth": []}],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["patient", "recommendedSports"],
                                "properties": {
                                    "patient": {
                                        "type": "string",
                                        "description": "ID of the patient"
                                    },
                                    "recommendedSports": {
                                        "type": "array",
                                        "description": "List of recommended sports",
                                        "items": {
                                            "type": "object",
                                            "required": ["name", "intensity", "benefits"],
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "description": "Name of the sport"
                                                },
                                                "intensity": {
                                                    "type": "string",
                                                    "enum": ["low", "medium", "high"],
                                                    "description": "Intensity level of the sport"
                                                },
                                                "benefits": {
                                                    "type": "string",
                                                    "description": "Health benefits of the sport"
                                                },
                                                "duration": {
                                                    "type": "string",
                                                    "description": "Recommended duration"
                                                },
                                                "frequency": {
                                                    "type": "string",
                                                    "description": "Recommended frequency"
                                                }
                                            }
                                        }
                                    },
                                    "notes": {
                                        "type": "array",
                                        "description": "Additional notes and recommendations",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "createdAt": {
                                        "type": "string",
                                        "format": "date-time",
                                        "description": "Creation timestamp"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sport recommendation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/sports/all": {
            "get": {
                "tags": ["Sports"],
                "summary": "Get all sport recommendations",
                "security": [{"bearerAuth": []}],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of all sport recommendations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "benefits": {
                                                        "type": "string"
                                                    },
                                                    "intensity": {
                                                        "type": "string",
                                                        "enum": ["low", "medium", "high"]
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/sports/get/{id}": {
            "get": {
                "tags": ["Sports"],
                "summary": "Get a sport recommendation by ID",
                "security": [{"bearerAuth": []}],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Sport recommendation ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sport recommendation details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "benefits": {
                                                    "type": "string"
                                                },
                                                "intensity": {
                                                    "type": "string",
                                                    "enum": ["low", "medium", "high"]
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sport recommendation not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/sports/update/{id}": {
            "put": {
                "tags": ["Sports"],
                "summary": "Update a sport recommendation",
                "security": [{"bearerAuth": []}],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Sport recommendation ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "description": "Updated name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "description": "Updated description"
                                    },
                                    "benefits": {
                                        "type": "string",
                                        "description": "Updated benefits"
                                    },
                                    "intensity": {
                                        "type": "string",
                                        "enum": ["low", "medium", "high"],
                                        "description": "Updated intensity level"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sport recommendation updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sport recommendation not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/sports/delete/{id}": {
            "delete": {
                "tags": ["Sports"],
                "summary": "Delete a sport recommendation",
                "security": [{"bearerAuth": []}],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Sport recommendation ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sport recommendation deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sport recommendation not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}